+++ CHANGELOG.md
@@ -44,6 +44,7 @@ script to proceed.
 - Download instance inventory as CSV.
 - Format PostgreSQL start time in dashboard.
 - Avoid modal for dashboard refresh error.
+- temBoard agent auto_configure.sh conditionnaly enable statements plugins.
 
 
 ## [7.11] - Unreleased